requirements.txt
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
factory-boy==2.8.1
Faker==0.7.7
Markdown==2.6.7
nose==1.3.7
psycopg2==2.6.2
python-dateutil==2.6.0
six==1.10.0
wheel==0.24.0
